requirements.txt
mcp>=1.2.0
garminconnect>=0.1.47
python-dotenv>=1.0.0